Removed _logit.h
