+++ scipy/special/_add_newdocs.py
@@ -4319,68 +4319,6 @@ add_newdoc("exp2",
            [ 1.41421356,  2.        ,  2.82842712]])
     """)
 
-add_newdoc('expit',
-    """
-    expit(x, out=None)
-
-    Expit (a.k.a. logistic sigmoid) ufunc for ndarrays.
-
-    The expit function, also known as the logistic sigmoid function, is
-    defined as ``expit(x) = 1/(1+exp(-x))``.  It is the inverse of the
-    logit function.
-
-    Parameters
-    ----------
-    x : ndarray
-        The ndarray to apply expit to element-wise.
-    out : ndarray, optional
-        Optional output array for the function values
-
-    Returns
-    -------
-    scalar or ndarray
-        An ndarray of the same shape as x. Its entries
-        are `expit` of the corresponding entry of x.
-
-    See Also
-    --------
-    logit
-
-    Notes
-    -----
-    As a ufunc expit takes a number of optional
-    keyword arguments. For more information
-    see `ufuncs <https://docs.scipy.org/doc/numpy/reference/ufuncs.html>`_
-
-    .. versionadded:: 0.10.0
-
-    Examples
-    --------
-    >>> import numpy as np
-    >>> from scipy.special import expit, logit
-
-    >>> expit([-np.inf, -1.5, 0, 1.5, np.inf])
-    array([ 0.        ,  0.18242552,  0.5       ,  0.81757448,  1.        ])
-
-    `logit` is the inverse of `expit`:
-
-    >>> logit(expit([-2.5, 0, 3.1, 5.0]))
-    array([-2.5,  0. ,  3.1,  5. ])
-
-    Plot expit(x) for x in [-6, 6]:
-
-    >>> import matplotlib.pyplot as plt
-    >>> x = np.linspace(-6, 6, 121)
-    >>> y = expit(x)
-    >>> plt.plot(x, y)
-    >>> plt.grid()
-    >>> plt.xlim(-6, 6)
-    >>> plt.xlabel('x')
-    >>> plt.title('expit(x)')
-    >>> plt.show()
-
-    """)
-
 add_newdoc("expm1",
     """
     expm1(x, out=None)
@@ -7989,135 +7927,6 @@ add_newdoc("log1p",
 add_newdoc("_log1pmx",
     """
     Internal function, do not use.
-    """)
-
-add_newdoc('log_expit',
-    """
-    log_expit(x, out=None)
-
-    Logarithm of the logistic sigmoid function.
-
-    The SciPy implementation of the logistic sigmoid function is
-    `scipy.special.expit`, so this function is called ``log_expit``.
-
-    The function is mathematically equivalent to ``log(expit(x))``, but
-    is formulated to avoid loss of precision for inputs with large
-    (positive or negative) magnitude.
-
-    Parameters
-    ----------
-    x : array_like
-        The values to apply ``log_expit`` to element-wise.
-    out : ndarray, optional
-        Optional output array for the function results
-
-    Returns
-    -------
-    out : scalar or ndarray
-        The computed values, an ndarray of the same shape as ``x``.
-
-    See Also
-    --------
-    expit
-
-    Notes
-    -----
-    As a ufunc, ``log_expit`` takes a number of optional keyword arguments.
-    For more information see
-    `ufuncs <https://docs.scipy.org/doc/numpy/reference/ufuncs.html>`_
-
-    .. versionadded:: 1.8.0
-
-    Examples
-    --------
-    >>> import numpy as np
-    >>> from scipy.special import log_expit, expit
-
-    >>> log_expit([-3.0, 0.25, 2.5, 5.0])
-    array([-3.04858735, -0.57593942, -0.07888973, -0.00671535])
-
-    Large negative values:
-
-    >>> log_expit([-100, -500, -1000])
-    array([ -100.,  -500., -1000.])
-
-    Note that ``expit(-1000)`` returns 0, so the naive implementation
-    ``log(expit(-1000))`` return ``-inf``.
-
-    Large positive values:
-
-    >>> log_expit([29, 120, 400])
-    array([-2.54366565e-013, -7.66764807e-053, -1.91516960e-174])
-
-    Compare that to the naive implementation:
-
-    >>> np.log(expit([29, 120, 400]))
-    array([-2.54463117e-13,  0.00000000e+00,  0.00000000e+00])
-
-    The first value is accurate to only 3 digits, and the larger inputs
-    lose all precision and return 0.
-    """)
-
-add_newdoc('logit',
-    """
-    logit(x, out=None)
-
-    Logit ufunc for ndarrays.
-
-    The logit function is defined as logit(p) = log(p/(1-p)).
-    Note that logit(0) = -inf, logit(1) = inf, and logit(p)
-    for p<0 or p>1 yields nan.
-
-    Parameters
-    ----------
-    x : ndarray
-        The ndarray to apply logit to element-wise.
-    out : ndarray, optional
-        Optional output array for the function results
-
-    Returns
-    -------
-    scalar or ndarray
-        An ndarray of the same shape as x. Its entries
-        are logit of the corresponding entry of x.
-
-    See Also
-    --------
-    expit
-
-    Notes
-    -----
-    As a ufunc logit takes a number of optional
-    keyword arguments. For more information
-    see `ufuncs <https://docs.scipy.org/doc/numpy/reference/ufuncs.html>`_
-
-    .. versionadded:: 0.10.0
-
-    Examples
-    --------
-    >>> import numpy as np
-    >>> from scipy.special import logit, expit
-
-    >>> logit([0, 0.25, 0.5, 0.75, 1])
-    array([       -inf, -1.09861229,  0.        ,  1.09861229,         inf])
-
-    `expit` is the inverse of `logit`:
-
-    >>> expit(logit([0.1, 0.75, 0.999]))
-    array([ 0.1  ,  0.75 ,  0.999])
-
-    Plot logit(x) for x in [0, 1]:
-
-    >>> import matplotlib.pyplot as plt
-    >>> x = np.linspace(0, 1, 501)
-    >>> y = logit(x)
-    >>> plt.plot(x, y)
-    >>> plt.grid()
-    >>> plt.ylim(-6, 6)
-    >>> plt.xlabel('x')
-    >>> plt.title('logit(x)')
-    >>> plt.show()
-
     """)
 
 add_newdoc("lpmv",
